requirements.txt
aiogram>=3.4,<4
aiosqlite>=0.19
python-dotenv>=1.0
aiohttp>=3.9
colorlog>=6.7
openpyxl>=3.1
pillow>=10.0
